requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2
django-filter==23.2
django-import-export==3.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
iniconfig==2.0.0
MarkupPy==1.14
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.26.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
toml==0.10.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
xlrd==2.0.1
xlwt==1.3.0